% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{autoapply}
\alias{autoapply}
\title{Autoapply a lapply to a function if the parameter is a list instead of a parameter.}
\usage{
autoapply(clist, fun, ...)
}
\arguments{
\item{clist}{Argument that is possibly a list}

\item{fun}{The function}

\item{...}{additional arguments}
}
\description{
Autoapply a lapply to a function if the parameter is a list instead of a parameter.
}

