% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check.model}
\alias{check.model}
\title{Check Model
Checks a test according to the model library to validate if it can be estimated or abort the current procedure}
\usage{
check.model(model, msg = "", error = T)
}
\arguments{
\item{model}{The model to check}

\item{msg}{Additional error message}

\item{error}{Optional, If false, this function wont throw a stop.}
}
\value{
No return, in case that the model is not valid throws a stop, if error is false, Only prints a message
}
\description{
Check Model
Checks a test according to the model library to validate if it can be estimated or abort the current procedure
}

