% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploratory.R
\name{Cronbach-Mesbah Curve}
\alias{Cronbach-Mesbah Curve}
\alias{alpha.c}
\title{Cronbach-Mesbah Curve}
\usage{
alpha.c(test)
}
\arguments{
\item{test}{a Dataframe that holds the test response data}
}
\value{
The number of items used to calculate the coefficient.

The maximum value of the alpha coefficient calculated at each step.

The item removed at each step.

Tue Cronbach-Mesbah curve plot.
}
\description{
To assess the unidimensionality of a set of items from alpha coefficient.
License :GPL (>= 2) Extracted from CMC_1.0 package
}
\details{
To construct the curve takes the next step by step:
1. The first step uses all items to compute alpha. 
2. One item is removed from the scale. The removed item is that which leaves the scale
with its maximum alpha value. If we remove a bad item, the alpha  coefficient  will
increase, whereas if we remove a good item alpha must decrease.  
3. This procedure is repeated until only  two items remain.
}
\references{
Cameletti, M. & Caviezel, V. (2010). Checking the unidimensionality in R
using the Cronbach-Mesbah curve.

Mesbah, M. (2010). Statistical quality of life. In "Method and Applications of Statistics in the 
Life and Health Sciences", N. BalakrishnanEd., Wiley, pp. 839-864.
}

