% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cronbach}
\alias{cronbach}
\title{Cronbach's alpha
Extacted from multilevel_2.5 package License: GPL (>= 2)}
\usage{
cronbach(items)
}
\arguments{
\item{items}{Dataframe that holds the test response data}
}
\value{
Cronbach's alpha for the test.
}
\description{
Cronbach's alpha
Extacted from multilevel_2.5 package License: GPL (>= 2)
}

