% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploratory.R
\name{Cronbachs alpha}
\alias{Cronbachs alpha}
\alias{alpha.cronbach}
\title{Alpha de Cronbach}
\usage{
cronbach(test)
}
\arguments{
\item{test}{a matrix or a Dataframe that holds the test response data}
}
\value{
Cronbach's alpha for the test.
}
\description{
Cronbach's alpha measures how correlated are the items in a test
License :GPL (>= 2) Extracted from multilevel_2.5 package
}
\details{
the coefficient is calculated alpha = (n/n-1)*(1 - (sum V_i/V_t))
where V_t is the variance of test scores and V_i is the variance of item scores.
It is desirable that the items are closely interrelated (coefficient near 1).
}
\references{
Cronbach L. J. (1951) Coefficient Alpha and the internal structure of tests. Psychometrika, 16,297-334
}

