% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{full.pattern.expand}
\alias{full.pattern.expand}
\title{full.pattern.expand}
\usage{
full.pattern.expand(pts, expandcol = ncol(pts))
}
\arguments{
\item{pts}{A matrix with frequencies in some column}

\item{expandcol}{A given frequency column, defaults to the last column}
}
\description{
Parameter expansion by a given column.
}

