% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{iapply}
\alias{iapply}
\title{Apply a function to the indices of a list}
\usage{
iapply(FUN, X, idx, ...)
}
\arguments{
\item{FUN}{a function}

\item{X}{a vector}

\item{idx}{a vector with the indices to consider from the list}
}
\description{
Apply a function to the indices of a list
}
\examples{
k=list(3,4,5,6)
idx=list(2,4)
iapply(function(x) x*x,k,idx)
}

