% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit.R
\name{iitem.hessian}
\alias{iitem.hessian}
\alias{item.hessian}
\title{item.hessian}
\usage{
item.hessian(test, zitem, th)
}
\arguments{
\item{test}{The original test}

\item{th}{The expanded latent traits}

\item{zitem.}{The item}
}
\value{
A matrix, the hessian
}

