% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{imapply}
\alias{imapply}
\title{Apply a function over indices of multiple lists and store in the original array
Warning, some functions need dots to be added to pass the additional arguments}
\usage{
imapply(X, FUN, idx, ...)
}
\description{
Apply a function over indices of multiple lists and store in the original array
Warning, some functions need dots to be added to pass the additional arguments
}
\examples{
k=list(3,4,5,6)
y=list(1,0,1,0)
idx=list(2,4)
p=mapply(function(x,y) x*y,k,y)
p
y=list(1,1,1,1)
imapply(p,function(x,y,...){x*y},idx,k,y)
}

