% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit.R
\name{ind.hessian}
\alias{ind.hessian}
\title{ind.hessian}
\usage{
ind.hessian(pattern, z, trait)
}
\arguments{
\item{pattern}{The original individual pattern}

\item{trait}{the individual trait}

\item{z.}{The item parameter list}
}
\value{
A numeric hessian
}

