% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory.R
\name{information}
\alias{information}
\title{Test or Item information}
\usage{
information(object, range, items = NULL, ...)
}
\arguments{
\item{object}{a matrix that contain the coefficient estimates}

\item{range}{a interval for which the test information should be computed.}

\item{items}{the items for which the information should be computed.}

\item{...}{Additional arguments to integrate}
}
\value{
TotalInfo the total amount of information.

RangeInfo the amount of information in the specified interval.

RangeProp the proportion of information in the specified interval.

Range the value of range argument

items the value of items argument
}
\description{
Computes the amount of test or item information.
License :GPL (>= 2) Adapted from CMC_1.0 package
}

