% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimreduction.R
\name{IPCA}
\alias{IPCA}
\title{IPCA}
\usage{
IPCA(data, h7 = 0.9, h2 = 0.2, hn = 10, angu = 15, clas = TRUE,
  reclas = TRUE)
}
\arguments{
\item{test:}{Dichotomic matrix data.}

\item{h7:}{Threshold to define the items not classificated}

\item{h2:}{Threshold to define the unidimensionality}

\item{hn:}{number of items in each cluster}

\item{angu:}{Threshol to define the angle of each cluster (it could change)}

\item{clas:}{If is TRUE then the classification is made, default is TRUE}

\item{reclas:}{If is TRUE then the reclassification is made, default is TRUE
If clas=TRUE and reclas=TRUE then the results are:}
}
\value{
n.submatrix: Total of submatrices Before and After reclassification

n.ipm.befrec: Number of items per matrix before reclassification. If clas=TRUE and reclas=FALSE or clas=FALSE and reclas=FALSE, then this result is NULL.

n.ipm.aftrec: Number of items per matrix after reclassification. If clas=TRUE and reclas=FALSE or clas=FALSE and reclas=FALSE, then this result is NULL.

n.ipm.clas: Number of items per matrix after classification. If clas=FALSE and reclas=TRUE or clas=FALSE and reclas=FALSE, then this result is NULL.

p.axes: Principal Axis of each submatrix

ipm: Matrix of items on each submatrix before classification. If clas=FALSE and reclas=TRUE or clas=FALSE and reclas=FALSE, then this result only show the first classification.

ipmc: Matrix of items on each submatrix after classification. If clas=FALSE and reclas=TRUE or clas=FALSE and reclas=FALSE, then this result is NULL.

DPM: Data of each submatrix (List object)

inc: Items not classified in a principle
}
\description{
Iterative Principal Components Analysis (IPCA)  Computes the Iterative components analysis to find dimensionality of a huge data matrix with entries 0 and 1.
}
\section{For textual data analysis is recomended to make ortogonalization of axes of each dimension to have interpretation of them.}{

}
\examples{

irtpp()
}
\author{
SICS Research Group, Universidad Nacional de Colombia \email{ammontenegrod@unal.edu.co}
}
\references{
Montenegro, A.M. & Ordo\~nez, M.F & P\'aez, S (2015)  A Novel Methodology Based on Item Responde Theory to Find Dimensionality and Emergent Information from Textual Data. \emph{Journal of Infometrix(Submited)}
}
\seealso{
\code{\link{z3_itemf}}, \code{\link{x2_itemf}}
}

