\name{IRTpp}
\alias{IRTpp}
\alias{IRTpp}
\docType{package}
\title{
IRTpp Item Response Theory implemented in cpp
}
\description{
Estimating item parameters and individual parameters under MIRT methodology
}
\details{
\tabular{ll}{
Package: \tab IRTpp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-19-03\cr
License: \tab MIT \cr
}
Use the function irtpp
}
\author{
SICS Research Team

Maintainer: Who to complain to <jcliberatol@unal.edu.co>
}
\references{
Andrade
}
\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
\dontrun{
irtpp()
}

}
