% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{irtpp.model}
\alias{irtpp.model}
\title{Return a valid model string for a representation of a model.}
\usage{
irtpp.model(model)
}
\arguments{
\item{model}{Model representation in a integer, string, character, or list with a model element}
}
\value{
model The valid string model for this model.
}
\description{
Return a valid model string for a representation of a model.
}

