% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{irtpp.p}
\alias{irtpp.p}
\title{Probability function for all models.}
\usage{
irtpp.p(model)
}
\arguments{
\item{model}{The model to calculate the probability to}
}
\description{
Probability function for all models.
}

