% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTpp-package.R
\docType{package}
\name{IRTpp}
\alias{IRTpp}
\alias{IRTpp-package}
\title{IRTpp : Item Response Theory Implemented in R and Cpp}
\description{
IRTpp supports the IRT and MIRT methodologies by implementing 
many IRT relevant algorithms for model exploration, calibration and 
fitting, IRTpp aims to be a high IRT performance software, using C++ where it is
needed.  IRTpp can be used not only on classical IRT applications but 
also on data mining. Particularly in textual analysis.
}
\details{
\tabular{ll}{
Package: \tab IRTpp\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-19-03\cr
License: \tab MIT \cr
}
}
\section{Getting Started}{

Get started with the IRTpp package browsing the index of this documentation
if you need help the vignettes should be helpful.


The IRTpp package allows you to use the IRT methodology for simulating, analyzing and scoring tests \cr
You can browse the package vignettes to get started.
}
\author{
SICS Research Team
}
\keyword{IRT}
\keyword{MIRT}
\keyword{Psychometry}

