% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itemfit.R
\name{itemfit}
\alias{itemfit}
\title{itemfit
evalua que tan bien se ajusta el modelo a los datos.}
\usage{
itemfit(model, z, patterns, theta)
}
\arguments{
\item{model}{el modelo implementado.}

\item{z}{Parametros de los items}

\item{patterns}{Patrones y frecuencias}

\item{theta}{Trazos por patron}
}
\value{
est Estadistica X² y P-valor
}
\description{
itemfit
evalua que tan bien se ajusta el modelo a los datos.
}

