% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itemfit.R
\name{itemFit2_p}
\alias{itemFit2_p}
\title{Z3 PERSONFIT
Calcula la estadística Z3}
\usage{
itemFit2_p(data, zita, patterns)
}
\arguments{
\item{data:}{datos}

\item{zita:}{Parametros de los items (con el c en todo R y el d en lugar del b)}

\item{patterns:}{Matriz con los patrones, las frecuencias y los trazos.}
}
\value{
Z3_personfit
}
\description{
Z3 PERSONFIT
Calcula la estadística Z3
}

