% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{loglik.f}
\alias{loglik.f}
\title{LogLikelihood of a IRT model for UIRT (Fast version)}
\usage{
loglik.f(test, traits, z)
}
\arguments{
\item{test}{A matrix of 0's and 1's}

\item{traits}{A vector with each individual parameter, or list of}

\item{z}{A list with the parameters a b and c specified by keys.
Each key must contain a vector of the item parameters for each parameter}
}
\description{
LogLikelihood of a IRT model for UIRT (Fast version)
}

