% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{loglik}
\alias{loglik}
\title{LogLikelihood of a IRT model}
\usage{
loglik(model, est)
}
\arguments{
\item{model}{The model to get the probability from}

\item{est}{The estimation containing the item parameters, test and individual parameters}
}
\description{
LogLikelihood of a IRT model
}

