% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itemfit.R
\name{orlando_itemf}
\alias{orlando_itemf}
\title{Orlando's statistic}
\usage{
orlando_itemf(patterns, G, zita, model)
}
\arguments{
\item{G}{number of quadrature Points}

\item{model}{type of model ( "1PL", 2PL", "3PL" )}

\item{patterns:}{matrix of patterns response, the frequency of each pattern and the latent traits}

\item{zita:}{matrix of estimations of the parameters of the items (discrimination,difficulty, guessing)}
}
\value{
Orlando's statistic, degrees of freedom and pvalue for each item
}
\description{
calculates the S-X2 values from Maria Orlando and David Thisen.
}
\examples{
## Simulates data "3PL" (5 items, default: 1000 individuals)
data=simulateTest("3PL",items=5)
data=data$test

## fit a model "3PL"
model=irtpp(data,"3PL")

## objects required  by "orlando_itemf()"
zita=parameter.matrix(model$zita)
it=individual.traits("3PL",zita,dataset = data,method = "EAP")
patterns=as.matrix(cbind(pattern.freqs(data,it),it$trait)) #patterns (patrones, frecuencia y trazos)
G = 61

## Orlando`s statistic

statistic=orlando_itemf(patterns=patterns,G=G,zita=zita,model="3PL")
statistic
}
\author{
SICS Research, National University of Colombia \email{ammontenegrod@unal.edu.co}
}
\references{
Orlando, M. & Thissen, D. (2000). Likelihood-based item fit indices for dichotomous item
response theory models. \emph{Applied Psychological Measurement, 24}, 50-64.
}
\seealso{
\code{\link{z3_itemf}}, \code{\link{x2_itemf}}
}

