% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory.R
\name{Parallel Analysis}
\alias{Parallel Analysis}
\alias{parallel}
\title{Parallel Analysis of a Correlation Matrix
License :GPL Extracted from nFactors_2.3.2 package}
\usage{
parallel(subject = 100, var = 10, rep = 100, cent = 0.05,
  quantile = cent, model = "components", sd = diag(1, var), ...)
}
\arguments{
\item{subject}{a number of subjects (default is 100)}

\item{var}{a number of variables (default is 10)}

\item{rep}{a number of replications of the correlation matrix (default is 100)}

\item{quantile}{a quantile of the distribution on which the decision is made}

\item{model}{a character "components" or "factors"}

\item{sd}{a vector of standard deviations of the simulated variables}

\item{...}{other parameters for the "mvtnorm" or "corr" functions}
}
\value{
eigen a Data frame containing the information on the distribution of the eigenvalues

eigen$mevpea Mean of the eigenvalues distribution

eigen$sevpea Standard deviation of the eigenvalues distribution

eigen$qevpea Quantile of the eigenvalues distribution

eigen$sqevpea Standard erro of the quantile of the eigenvalues distribution

subject a number of subjects

variables a number of variables

centile Selected quantile
}
\description{
Parallel Analysis of a Correlation Matrix
License :GPL Extracted from nFactors_2.3.2 package
}

