% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{Parameter Initialization}
\alias{Parameter Initialization}
\alias{parameters.initialize}
\title{parameters.initialize}
\usage{
parameters.initialize(dataset, model, dims, method = "DEFAULT",
  red.items = T)
}
\arguments{
\item{dataset}{The dataset to calculate the initial parameters. A matrix of 0's  and 1's}

\item{model}{The model with dimensions.}

\item{dims}{To set the dimensions of use in the test.}

\item{method.}{Optional, "PCA" for multidimensional and "ANDRADE" for unidimensional are teh current implementations}

\item{red.items.}{Optional, default true. Reduces the dataset to the dataset without trash items.s}
}
\value{
Initial values for a estimation on the dataset and the model.
}
\author{
SICS Research Group
}

