% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter.list}
\alias{parameter.list}
\title{Returns a parameter list from a parameter matrix.}
\usage{
parameter.list(pars, model = "3PL", dpar = NULL, cp = NULL)
}
\arguments{
\item{pars}{The parameter matrix}

\item{model}{The model whose the matrix refers to.}
}
\description{
Returns a parameter list from a parameter matrix.
}

