% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialization.R
\name{ParameterInitialization}
\alias{ParameterInitialization}
\alias{parameters.initialize}
\title{parameters.initialize}
\usage{
parameters.initialize(dataset, model, method = "DEFAULT", red.items = T)
}
\arguments{
\item{dataset}{The dataset to calculate the initial parameters. A matrix of 0's  and 1's}

\item{model}{The model with dimensions.}

\item{method.}{Optional, "PCA" for multidimensional and "ANDRADE" for unidimensional are teh current implementations}

\item{red.items.}{Optional, default true. Reduces the dataset to the dataset without trash items.s}
}
\value{
Initial values for a estimation on the dataset and the model.
}
\description{
parameters.initialize
}
\author{
SICS Research Group
}

