% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern.expand}
\alias{pattern.expand}
\title{pattern.expand}
\usage{
pattern.expand(pts)
}
\arguments{
\item{pts}{The pattern matrix with frequencies}
}
\value{
test. The original test of the patterns.
}
\description{
Expands a pattern frequency matrix by the frequencies in the last column. Inverse of pattern.freqs
}

