% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern.freqs}
\alias{pattern.freqs}
\title{pattern.freqs}
\usage{
pattern.freqs(data, traitobj = NULL, onlyTheta = T)
}
\arguments{
\item{test}{The data matrix}
}
\value{
Patterns and frequencies in a matrix , the last column are the frequencies
}
\description{
Calculates the pattern frequencies given a dataset.
}

