% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory.R
\name{Plot Parallel}
\alias{Plot Parallel}
\alias{plotParallel}
\title{Plot a Parallel Analysis
License: GPL Extracted from nFactors_2.3.2 package}
\usage{
plotParallel(parallel, eig = NA, x = eig, model = "components",
  legend = TRUE, ylab = "Eigenvalues", xlab = "Components",
  main = "Parallel Analysis", ...)
}
\arguments{
\item{parallel}{a vector of the results of parallel analysis}

\item{eig}{eigenvalues to analyse}

\item{x}{a vector of eigenvalues, a matrix of correlations or a data.frame}

\item{model}{a character "components" or "factors"}

\item{...}{other graphics parameters}
}
\description{
Plot a Parallel Analysis
License: GPL Extracted from nFactors_2.3.2 package
}

