% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{print.sentence}
\alias{print.sentence}
\title{Print Sentence, Helper function
Prints different strings concatenating them with a " "}
\usage{
\method{print}{sentence}(..., sep = " ", verbose = T)
}
\arguments{
\item{...}{The strings to print.}

\item{sep}{The separator to pass}
}
\description{
Print Sentence, Helper function
Prints different strings concatenating them with a " "
}

