% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{prob.2pl}
\alias{prob.2pl}
\title{2PL fast probability function
The probability function in the 2PL model}
\usage{
prob.2pl(z, theta)
}
\arguments{
\item{z.}{List containing the item parameters a and d}

\item{theta.}{Vector, contains the latent traits of the individual}
}
\description{
2PL fast probability function
The probability function in the 2PL model
}

