% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{prob.3pl}
\alias{prob.3pl}
\title{3PL fast probability function
The probability function in the 3PL models}
\usage{
prob.3pl(z, theta)
}
\arguments{
\item{z.}{List containing the item parameters a, d and c.}

\item{theta.}{Vector, contains the latent traits of the individual}
}
\description{
3PL fast probability function
The probability function in the 3PL models
}

