% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{prob}
\alias{prob}
\title{prob.}
\usage{
prob(theta, a, d, c)
}
\arguments{
\item{theta:}{eltrazo latente multidimensional}

\item{a:}{parametro de discrimnacion en el caso multidimensional}

\item{d:}{parametro d en el caso multidimensional}

\item{c:}{parametro de azar en el caso multidimensional}
}
\value{
la probabilidad
}
\description{
calcula la probabilidad en el caso multidimensional
}
\author{
Juan David Cortés
}

