% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{probability.1pl}
\alias{probability.1pl}
\title{2PL probability function
The probability function in the 2PL model.}
\usage{
probability.1pl(z, b = z$b, theta)
}
\arguments{
\item{z}{Optional. A list with the parameter b specified by keys.}

\item{b}{The difficulty parameter. (Optional if d is specified)}

\item{theta}{The subject's latent trait.}
}
\description{
2PL probability function
The probability function in the 2PL model.
}

