% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{probability.2pl}
\alias{probability.2pl}
\title{2PL probability function
The probability function in the 3PL model.}
\usage{
probability.2pl(z, a = z$a, b = z$b, theta, d = -a * b)
}
\arguments{
\item{z}{Optional. A list with the parameters a and b specified by keys.}

\item{a}{The discrimination parameter}

\item{b}{The difficulty parameter. (Optional if d is specified)}

\item{theta}{The subject's latent trait.}

\item{d}{Optional. Overrides the b parameter, it is equal to -a*b. Used in some functions.}
}
\description{
2PL probability function
The probability function in the 3PL model.
}

