% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/models.R
\name{probability.3pl}
\alias{probability.3pl}
\title{3PL probability function
The probability function in the 3PL model.}
\usage{
probability.3pl(z, a = z$a, b = z$b, c = z$c, theta, d = -a * b,
  cp = NULL)
}
\arguments{
\item{z}{Optional. A list with the parameters a b and c specified by keys.}

\item{a}{The discrimination parameter}

\item{b}{The difficulty parameter. (Optional if d is specified)}

\item{c}{The guessing parameter}

\item{theta}{The subject's latent trait.}

\item{d}{Optional. Overrides the b parameter, it is equal to -a*b. Used in some functions.}

\item{cp}{Optional. Overrides the c parameter, it is logit(c)}
}
\description{
3PL probability function
The probability function in the 3PL model.
}

