% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{search.index}
\alias{search.index}
\title{Return the elements of a list on some indices}
\usage{
search.index(x, idx)
}
\description{
Return the elements of a list on some indices
}
\examples{
k=list(3,4,5,6)
search.index(k,c(2,3))
}

