\name{IRTpp-package}
\alias{IRTpp-package}
\alias{IRTpp}
\docType{package}
\title{
What the package does (short line)
}
\description{
IRTpp is an implementation of Item Response Theory for R , written in Cpp
IRTpp is fast and robust.
}
\details{
\tabular{ll}{
Package: \tab IRTpp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-16\cr
License: \tab GPL (>= 2)\cr
}
Use the function IRTpp
\author{
SICS Software team , and SICS research group

Maintainers
Francisco Ospina <fjospinas@unal.edu.co>
Juan Liberato <jcliberatol@unal.edu.co>
Cristian Sandoval <cesandovalp@unal.edu.co>

And the rest of SICS software team.

}
\references{
~~ Reckase M.D. Multidimensional Item Response theory ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
