% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{Simulate test Multidimensional}
\alias{Simulate test Multidimensional}
\alias{simulateTestMD}
\title{Simulate test Multidimensional}
\usage{
simulateTestMD(items = 10, individuals = 1000, dims = 3, clusters = 4,
  seed = 10, z = NULL, repetition = NULL)
}
\arguments{
\item{items.}{The items to simulate}

\item{individuals.}{The individuals that answer the test.}

\item{dims.}{The dimensions of the test.}

\item{seed.}{A seed for reproducibility.}

\item{z.}{A parameter list to feed the function.}

\item{repetition.}{A number of repetition to simulate with the same item parameters different tests.}
}
\value{
This function returns the following data in a named list :
\itemize{
  \item test : The simulated test
  \item z : A IRT parameter list.
  \item clusters : The list of cluster sizes per cluster.
  \item direction : The directions in the dimensions for each cluster.
  \item clustinit : The principal item of each cluster
  \item clusterlist : The range list of the clusters.
}
}
\description{
Simulate test Multidimensional
}
\author{
SICS Research team
}

