% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{SimulateItemParameters}
\alias{SimulateItemParameters}
\alias{simulateItemParameters}
\title{Simulates item parameters depending on a model}
\usage{
simulateItemParameters(items, model, dims = 1, boundaries = NULL)
}
\arguments{
\item{items}{, Number of items to generate}

\item{model}{A string with the model to simulate, please refer to the model documentation in irtpp documentation.}

\item{dims}{Optional. The number of dimensions to simulate in the test if the model is multidimensional}

\item{boundaries}{Optional. The kind of boundaries that are specified for the parameters.}
}
\description{
Simulates item parameters for UIRT models.
}

