% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/test.R
\name{simulateTest}
\alias{simulateTest}
\title{Simulate a test according to a model.}
\usage{
simulateTest(model, items, individuals, independent = TRUE, reps = 1,
  dims = 1, boundaries = NULL, generated = TRUE, itempars = NULL,
  seed = NULL)
}
\arguments{
\item{model}{A string with the model to simulate, please refer to the model documentation in irtpp documentation.}

\item{items}{the number of items to simulate}

\item{individuals}{the number of individuals to simulate}

\item{independent}{Set this to false if all the individuals used to simulate the test must be the same}

\item{reps}{The number of tests to generate with this settings}

\item{dims}{Optional. The number of dimensions to simulate in the test if the model is multidimensional}

\item{boundaries}{Optional. The kind of boundaries that are specified for the parameters.}

\item{itempars}{Optional. Item parameters to be used in the simulation. When the parameters are not generated, the item parameters must be specified.}

\item{seed}{Optional. Seed to use to generate all the data}
}
\value{
A data frame with the simulated dataset
}
\description{
Example \eqn{a + b^2}
}
\author{
Juan Liberato
}

