% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatePoly.R
\name{Simulation of test polytomous}
\alias{SimulateTestPoly}
\alias{Simulation of test polytomous}
\title{Simulate Test polytomous}
\usage{
SimulateTestPoly(n, thetas, model = c("gpcm", "grm"))
}
\arguments{
\item{n}{a value indicating the number of patterns to simulate}

\item{thetas}{a list containing the parameters of the items where the former 
are the values for the difficulty and the last value is that of discrimination}

\item{model}{a value indicating the model to use}
}
\value{
data the polytomous test according to the parameters established
}
\description{
Politomous test simulates according to models GRM and GPCM
}

