% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exploratory.R
\name{Test or Item information}
\alias{Test or Item information}
\alias{information}
\title{Test or Item information}
\usage{
information(object, range, items = NULL, ...)
}
\arguments{
\item{object}{a matrix  (con los coeficientes estimados)}

\item{range}{a interval for which the test information should be computed.}

\item{items}{the items for which the information shoulb be computed.}
}
\value{
TotalInfo the total amount of information.

RangeInfo the amount of information in the specified interval.

RangeProp the proportion of information in the specified interval.

Range the value of range argument

items the value of items argument
}
\description{
Computes the amount of test or item information. 
License :GPL (>= 2) Adapted from ltm_1.0 package
}
\details{
The amount of information is computed as the area under the Item or Test
Information Curve.
}
\references{
Reckase, M. (2009). Multidimensional item response theory. New York: Springer.

Baker, F. B., & Kim, S. H. (Eds.). (2004). Item response theory: Parameter estimation techniques. CRC Press.
}

