% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit.R
\name{test.hessian}
\alias{test.hessian}
\title{test.hessian}
\usage{
test.hessian(test, z, th, byitems = T, byinds = T)
}
\arguments{
\item{test}{The original test}

\item{th}{The expanded latent traits}

\item{byitems}{Calculate the hessian by items}

\item{z.}{The item parameter list}

\item{byinds.}{Calculate by individuals}
}
\value{
A list with $itemhessian and $indhessian
}

