% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test.plot}
\alias{test.plot}
\title{test.plot}
\usage{
test.plot(z, i = NULL)
}
\arguments{
\item{z}{: Item parameter list}

\item{i}{: Optional, item to plot.}
}
\value{
Void, draws a plot of the ICC or the test's ICC's
}
\description{
plots a test or an item.
}
\author{
Juan Liberato
}

