% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{testmulti}
\alias{testmulti}
\title{testmulti}
\usage{
testmulti(nitems, ninds, dim, model)
}
\arguments{
\item{nitems:}{numero de items}

\item{ninds:}{numero de individuos}

\item{dim:}{dimension del trazo}

\item{model:}{1pl, 2pl o 3pl}
}
\value{
el test y los parámeros poblacionales con los que se simuló
}
\description{
simula un test multidimensional
}
\author{
Juan David Cortés
}

