% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{ua}
\alias{ua}
\title{Undefined assignment, Helper function}
\usage{
ua(var, val)
}
\arguments{
\item{var}{, The variable to test}

\item{val}{, The value to return if the tested variables is NULL}
}
\value{
Returns the value of the tested variable if it is not NULL, otherwise, returns the default
}
\description{
Undefined assignment, Helper function
}

