% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory.R
\name{Unidimensional test}
\alias{Unidimensional test}
\alias{unidimTest}
\title{Test of Unidimensionality using Modified Parallel Analysis
License: GPL Extracted from ltm_1.0. package}
\usage{
unidimTest(object, data, thetas, IRT = TRUE, z.vals = NULL, B = 100, ...)
}
\arguments{
\item{object}{an object from either class ltm, rasch or tpm.}

\item{data}{a matrix or a data.frame; used if object is missing.}

\item{thetas}{a matrix with IRT model parameter values to be used in rmvlogis; used if object is missing.}

\item{IRT}{}

\item{z.vals}{}

\item{B}{a number of samples for the Monte Carlo procedure}
}
\description{
Test of Unidimensionality using Modified Parallel Analysis
License: GPL Extracted from ltm_1.0. package
}

