% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{z.item}
\alias{z.item}
\title{z.item}
\usage{
z.item(z, i)
}
\arguments{
\item{i}{Item or item vector to get.}
}
\value{
The parameter list with the items specified.
}
\description{
Reads of a parameter list by item or items.
}
\author{
Juan Liberato
}

