% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit.R
\name{z3_itemf}
\alias{z3_itemf}
\title{Z3 ITEMFIT
Calcula la estad??stica Z3}
\usage{
z3_itemf(data, zita, patterns)
}
\arguments{
\item{data:}{datos}

\item{zita:}{Parametros de los items (con el c en todo R y el d en lugar del b)}

\item{patterns:}{Matriz con los patrones, las frecuencias y los trazos.}
}
\value{
Z3_itemfit
}
\description{
Z3 ITEMFIT
Calcula la estad??stica Z3
}

