% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory.R
\name{Biserial coefficient}
\alias{Biserial coefficient}
\alias{biserial.cor}
\title{Point-Biserial correlation coefficient}
\usage{
biserial.cor(x, y, use = c("all.obs", "complete.obs"), level = 1)
}
\arguments{
\item{x}{a numeric vector representing the continuous variable.}

\item{y}{a numeric vector representing the dichotomous variable.}

\item{use}{Is a option for the use of missing values.}

\item{level}{which level of y to use.}
}
\value{
the value of the point-biserial correlation.
}
\description{
Point-Biserial correlation coefficient is a correlation coefficient
used when one variable is continuous and the other variable is dichotomous.
License :GPL Adapted from ltm_1.0 package
}

