% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseCells}
\alias{collapseCells}
\title{Corrige fr. esperadas muy pequeñas para la estadística de orlando}
\usage{
collapseCells(O, E, mincell = 1)
}
\arguments{
\item{O, }{es una lista (de longitud nitems) con fr. observadas de respuesta incorrectas y correctas para cada item}

\item{E, }{es una lista (de longitud nitems) con fr. esperadas de respuesta incorrectas y correctas para cada item}

\item{mincell, }{es el minimo numero esperado de personas con score k que contestan o no al item j}
}
\description{
Corrige fr. esperadas muy pequeñas para la estadística de orlando
}

