% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory.R
\name{Cronbach-Mesbah Curve}
\alias{Cronbach-Mesbah Curve}
\alias{alpha.curve}
\title{Cronbach-Mesbah Curve}
\usage{
alpha.curve(x)
}
\arguments{
\item{x}{a Dataframe that holds the test response data}
}
\value{
The number of items used to calculate the coefficient.

The maximum value of the alpha coefficient calculated at each step.

The item removed at each step.

Tue Cronbach-Mesbah curve plot.
}
\description{
Point-Biserial correlation coefficient is a correlation coefficient
used when one variable is continuous and the other variable is dichotomous.
License :GPL (>= 2) Extracted from CMC_1.0 package
}
\details{
aaa
}

