% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory.R
\name{Cronbachs alpha}
\alias{Cronbachs alpha}
\alias{cronbach}
\title{Alpha de Cronbach}
\usage{
cronbach(test)
}
\arguments{
\item{test}{a matrix or a Dataframe that holds the test response data}
}
\value{
Cronbach's alpha for the test.
}
\description{
Cronbach's alpha measures how correlated are the items in a test
License :GPL (>= 2) Extacted from multilevel_2.5 package
}

