% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/funcion_testmulti.R
\name{fix.items}
\alias{fix.items}
\title{fix.items
establece que items fijar en el modelo multi-unidimensional}
\usage{
fix.items(TEST, DIMTEST)
}
\arguments{
\item{TEST:}{el test que se quiere modelar}

\item{DIMTEST:}{un vector que indique que items corresponden a que dimension}
}
\description{
fix.items
establece que items fijar en el modelo multi-unidimensional
}

