% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/irtpp.R
\name{individual.traits}
\alias{individual.traits}
\title{Estimate the latent traits of the individuals in a test with some given item parameters}
\usage{
individual.traits(dataset, model, itempars)
}
\arguments{
\item{dataset}{The matrix with the responses from the individuals}

\item{model}{The model used to calibrate the parameters}

\item{itempars}{The item parameters for the model.}
}
\description{
Estimate the latent traits of the individuals in a test with some given item parameters
}

