% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/irtpp.R
\name{individual.traits}
\alias{individual.traits}
\title{Estimate the latent traits of the individuals in a test with some given item parameters}
\usage{
individual.traits(model, itempars, method, dataset = NULL, filename = NULL,
  output = NULL, probability_matrix = NULL)
}
\arguments{
\item{model}{The model used to calibrate the parameters}

\item{itempars}{The item parameters for the model.}

\item{method}{The method to estimate traits}

\item{dataset}{The matrix with the responses from the individuals}
}
\value{
A list with the patterns and the estimated latent traits
}
\description{
Estimate the latent traits of the individuals in a test with some given item parameters
}

