% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/irtpp.R
\name{irtpp}
\alias{irtpp}
\title{Estimate a test item parameters according to Item Response Theory.}
\usage{
irtpp(dataset = NULL, model, initialvalues = NULL, filename = NULL,
  output = NULL)
}
\arguments{
\item{dataset}{The matrix with the responses from the individuals}

\item{model}{The model used to calibrate the parameters}

\item{initialvalues}{The matrix with the initial values for the optimization process}

\item{filename}{Optional argument specifying a CSV file to read instead of a dataset in memory}

\item{output}{Optional. Additonal arguments that need to be documented by cristian}
}
\value{
The item parameters in a matrix.
}
\description{
Estimate a test item parameters according to Item Response Theory.
}

