% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IRTpp-package.R
\docType{package}
\name{IRTpp}
\alias{IRTpp}
\alias{IRTpp-package}
\title{IRTpp : Item Response Theory Implemented in R and Cpp}
\description{
IRTpp supports the IRT and MIRT methodologies by implementing
many IRT relevant algorithms for model exploration, calibration and fitting, IRTpp
aims to be extremely performant using Cpp where it is needed, focusing in extending
IRT not only to classical applications but also in new applications which sometimes
need highly performant implementations.
}
\details{
\tabular{ll}{
Package: \tab IRTpp\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-19-03\cr
License: \tab MIT \cr
}
}
\section{Getting Started}{

Get started with the IRTpp package browsing the index of this documentation
if you need help the vignettes should be helpful.
}

\section{Getting Started}{

The IRTpp package allows you to use the IRT methodology for simulating, analyzing and scoring tests \cr
You can browse the package vignettes to get started. \href{vignettes/simulateTest.html}{Sim test}
\href{help/library/IRTpp/doc/index.html}{Sim test2}
help/library/IRTpp/doc/index.html
\href{/help/library/IRTpp/doc/index.html}{Sim test3}
}
\author{
SICS Research Team
}
\keyword{IRT}
\keyword{MIRT}
\keyword{Psychometry}

