% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/irtpp.R
\name{irtpp}
\alias{irtpp}
\title{Estimate a test item parameters according to Item Response Theory.}
\usage{
irtpp(dataset, model)
}
\arguments{
\item{dataset}{The dataset to estimate}
}
\description{
Estimate a test item parameters according to Item Response Theory.
}

