% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itemfit.R
\name{itemfit}
\alias{itemfit}
\title{itemfit
evalua que tan bien se ajusta el modelo a los datos.}
\usage{
itemfit(a)
}
\arguments{
\item{a}{el modelo implementado.}
}
\description{
itemfit
evalua que tan bien se ajusta el modelo a los datos.
}

