% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{loglik}
\alias{loglik}
\title{LogLikelihood of a IRT model for UIRT}
\usage{
loglik(test, traits, z)
}
\arguments{
\item{test}{A matrix of 0's and 1's}

\item{traits}{A vector with each individual parameter, or list of}

\item{z}{A list with the parameters a b and c specified by keys.
Each key must contain a vector of the item parameters for each parameter}
}
\description{
LogLikelihood of a IRT model for UIRT
}

