% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/orlando.R
\name{orlando_itemf}
\alias{orlando_itemf}
\title{Orlando
Calcula la estadística de Orlando}
\usage{
orlando_itemf(patterns, G, zita, FUN)
}
\arguments{
\item{G}{numero de puntos de cuadratura}

\item{FUN}{: funcion de representante de cada grupo}

\item{patterns:}{matriz con los patrones de respuesta, las frecuencias y los trazos}

\item{zita:}{MATRIZ de estimaciones de parametros de los items (a,b,c)}
}
\value{
estadística de Orlando
}
\description{
Orlando
Calcula la estadística de Orlando
}

