% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/itemfit.R
\name{orlando_itemf}
\alias{orlando_itemf}
\title{Orlando's statistic}
\usage{
orlando_itemf(patterns, G, zita, model)
}
\arguments{
\item{G}{number of quadrature Points}

\item{model}{type of model ( "1PL", 2PL", "3PL" )}

\item{patterns:}{matrix of patterns response, the frequency of each pattern and the latent traits}

\item{zita:}{matrix of estimations of the parameters of the items (discrimination,difficulty, guessing)}
}
\value{
Orlando's statistic, degrees of freedom and pvalue for each item
}
\description{
calculates the S-X2 values from Maria Orlando and David Thisen.
}
\examples{
irtpp()
}
\author{
SICS Research, National University of Colombia \email{ammontenegrod@unal.edu.co}
}
\references{
Orlando, M. & Thissen, D. (2000). Likelihood-based item fit indices for dichotomous item
response theory models. \emph{Applied Psychological Measurement, 24}, 50-64.
}
\seealso{
\code{\link{z3_itemf}}, \code{\link{x2_itemf}}
}

