% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{s_ss}
\alias{s_ss}
\title{Verosimilitudes de los scores clasicos para calcular fr. esperadas en Orlando
Calcula las veorimilitudes}
\usage{
s_ss(pr, nitems, G)
}
\arguments{
\item{nitems, }{cantidad de items considerados en la matriz de probabilidad}

\item{pr:}{matriz de probabilidad}
}
\description{
Verosimilitudes de los scores clasicos para calcular fr. esperadas en Orlando
Calcula las veorimilitudes
}

