% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulateTest}
\alias{simulateTest}
\title{SimulateTest
Simulates a test according to a irt model}
\usage{
simulateTest(model = "2PL", items = 10, latentTraits = NULL,
  individuals = 1000, boundaries = NULL, dims = 1, itempars = NULL,
  verbose = F, threshold = 0, seed = NULL)
}
\value{
List with model, seed item pars, and the test simulated.
}
\description{
Simulates a test according to a IRT Model
}
\author{
Juan Liberato
}

