% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exploratory.R
\name{Test or Item information}
\alias{Test or Item information}
\alias{information}
\title{Test or Item information
License :GPL (>= 2) Extracted from CMC_1.0 package}
\usage{
information(object, range, items = NULL, ...)
}
\arguments{
\item{object}{an object from either class ltm, rasch or tpm.}

\item{range}{a interval for which the test information should be computed.}

\item{items}{the items for which the information shoulb be computed.}
}
\value{
TotalInfo the total amount of information.

RangeInfo the amount of information in the specified interval.

RangeProp the proportion of information in the specified interval.

Range the value of range argument

items the value of items argument
}
\description{
Test or Item information
License :GPL (>= 2) Extracted from CMC_1.0 package
}

